requirements.txt
mysql-connector-python==8.0.23
protobuf==3.17.3
setuptools==44.1.1
six==1.16.0
wxPython==4.0.0a1